requirements.txt
fastapi
flask
firebase_admin
uvicorn[standard]
line-bot-sdk==2.0.1
google.generativeai
grpcio
pillow
pytesseract
requests
git+https://github.com/ozgur/python-firebase
git+https://github.com/madmaze/pytesseract